refactor: structured the javascript code but needs to be tested
