+++ src/constants/assets.constant.js
@@ -1,0 +1,40 @@
+const BASE_MODEL = {
+	model_url: 'https://nbizksjfzehbiwmcipep.supabase.co/storage/v1/object/public/model/idle.glb',
+	// model_url: 'https://nbizksjfzehbiwmcipep.supabase.co/storage/v1/object/public/model/Idle_Head.glb',
+	animation: 'idle',
+};
+const ANIMATION_LIST = [
+	{
+		model_url: 'https://nbizksjfzehbiwmcipep.supabase.co/storage/v1/object/public/model/celebration.glb',
+		animation: 'celebration',
+	},
+	{
+		model_url: 'https://nbizksjfzehbiwmcipep.supabase.co/storage/v1/object/public/model/disappointed.glb',
+		animation: 'disappointed',
+	},
+	{
+		model_url: 'https://nbizksjfzehbiwmcipep.supabase.co/storage/v1/object/public/model/no_no.glb',
+		animation: 'no_no',
+	},
+	{
+		model_url: 'https://nbizksjfzehbiwmcipep.supabase.co/storage/v1/object/public/model/offer.glb',
+		animation: 'offer',
+	},
+	{
+		model_url: 'https://nbizksjfzehbiwmcipep.supabase.co/storage/v1/object/public/model/thumbs_up.glb',
+		animation: 'thumbs_up',
+	},
+	{
+		model_url: 'https://nbizksjfzehbiwmcipep.supabase.co/storage/v1/object/public/model/two_hand_wave.glb',
+		animation: 'two_hand_wave',
+	},
+	{
+		model_url: 'https://nbizksjfzehbiwmcipep.supabase.co/storage/v1/object/public/model/wave.glb',
+		animation: 'wave',
+	},
+];
+const MODEL_TEXTURE = 'https://nbizksjfzehbiwmcipep.supabase.co/storage/v1/object/public/model/base%20colour%20(1).png';
+const AUDIO_PATH = 'https://nbizksjfzehbiwmcipep.supabase.co/storage/v1/object/public/model/notification.mp3';
+const CLOSE_ICON = 'https://nbizksjfzehbiwmcipep.supabase.co/storage/v1/object/public/model/X%20Close%20Icon.png';
+
+module.exports = { BASE_MODEL, ANIMATION_LIST, MODEL_TEXTURE, AUDIO_PATH, CLOSE_ICON };
