+++ src/constants/definitions.js
@@ -1,0 +1,6 @@
+const CHATBOT_PAGE = 'https://saas-dashboard-henna.vercel.app';
+// const CHATBOT_PAGE = 'http://localhost:3000';
+const ENDPOINT = 'https://node-service-1e6u.onrender.com';
+// const ENDPOINT = 'http://localhost:3001'
+
+module.exports = { CHATBOT_PAGE, ENDPOINT };
